feat: add authentication

+++ graphql/resolvers/booking.js
@@ -1,9 +1,12 @@
-const Event = require('../../models/event');
-const Booking = require('../../models/booking');
-const {transformBooking, transformEvent } = require('./merge');
+const Event = require("../../models/event");
+const Booking = require("../../models/booking");
+const { transformBooking, transformEvent } = require("./merge");
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     try {
       const bookings = await Booking.find();
       return bookings.map(booking => {
@@ -13,18 +16,24 @@ module.exports = {
       throw err;
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
-      user: '6457d7cb260219574fd26c78',
+      user: req.userId,
       event: fetchedEvent
     });
     const result = await booking.save();
     return transformBooking(result);
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unauthenticated!");
+    }
     try {
-      const booking = await Booking.findById(args.bookingId).populate('event');
+      const booking = await Booking.findById(args.bookingId).populate("event");
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
       return event;
@@ -33,3 +42,4 @@ module.exports = {
     }
   }
 };
+
